Restored use of OpenGL in mainDisplay by shifting all trajectories to a local origin. The origin is displayed as text in ECEF coordinates.
In matlabGA1, fixed a bug in the computation of how many extension blocks to add to a dynamic model. Also, cost is now normalized by the total number of blocks and edges in the objective.

Fixed bug in boundedMarkov where some configuration variables still referred to dynamicModelStub (which has been depricated).